Factor out node-finding routines into their own file.
